refactor(wal): add error handlings
